101-print_number.c

+++ more_functions_nested_loops/main.h
@@ -13,6 +13,6 @@ void print_line(int n);
 void print_diagonal(int n);
 void print_square(int size);
 void print_triangle(int size);
-void print_number(long int n);
+void print_number(int n);
 
 #endif
